Tried to make an Interface out of Communication.cpp
Seems to be a problem with declaring the servIP address which should
point to localhost.
When it's fixed the idea of this interface is that the CreateSocket and
CloseSocket are fixed functions and the SendAndReceive function has to
be implemented for each class that inherits from this Interface. There
should be a way with less duplication but this will do for now.