Made a SluiceFactory and added functions to Door
SluiceFactory has an empty constructor and has to be controlled by the
SluiceWatch class this calls for a little change in our Design.

Doors now have a open and close function in this function the specific
calls have to be made to open and close the door. These functions won't
be fully inherited by DoorLock en DoorMotor but will require another
implementation. To keep all classes reusable the door takes care of its
lock and motorcontrol so that the sluice never has to know about the
type of door he is controlling.